requirements.txt
django~=3.2.13
djangorestframework~=3.13.1
djangorestframework-api-key
markdown~=3.3.7
django-mdeditor~=0.1.20
pillow
python-dotenv